+++ packages/react-magma-dom/src/components/Input/Input.stories.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Input, InputProps } from '.';
 import { InputIconPosition, InputSize, InputType } from '../InputBase';
 import { Story, Meta } from '@storybook/react/types-6-0';
-import { NotificationsIcon } from 'react-magma-icons';
+import { HelpIcon, NotificationsIcon } from 'react-magma-icons';
 import { Card, CardBody } from '../Card';
 import { magma } from '../../theme/magma';
 
@@ -45,6 +45,21 @@ Large.args = {
 export const File = Template.bind({});
 File.args = {
   type: InputType.file,
+};
+
+export const IconTop = Template.bind({});
+IconTop.args = {
+  ...Default.args,
+  icon: <HelpIcon />,
+  iconPosition: InputIconPosition.top,
+};
+
+export const IconTopLarge = Template.bind({});
+IconTopLarge.args = {
+  ...Default.args,
+  icon: <HelpIcon />,
+  inputSize: InputSize.large,
+  iconPosition: InputIconPosition.top,
 };
 
 export const IconLeft = Template.bind({});