+++ packages/react-magma-dom/src/components/Tooltip/index.tsx
@@ -33,6 +33,10 @@ export interface TooltipProps extends React.HTMLAttributes<HTMLDivElement> {
   content: React.ReactNode;
   isInverse?: boolean;
   /**
+   * Override the default opening of the tooltip on hover/focus to remain open
+   */
+  open?: boolean;
+  /**
    * Position the tooltip appears in relation to its trigger
    */
   position?: TooltipPosition;
@@ -48,12 +52,15 @@ export interface ITooltipState {
   isVisible?: boolean;
 }
 
-const ToolTipContainer = styled.div`
+const TooltipContainer = styled.div`
   display: inline;
   pointer-events: auto;
 `;
 
-const ToolTipArrow = styled.span<{ position?: any; isInverse?: boolean }>`
+export const TooltipArrow = styled.span<{
+  position?: any;
+  isInverse?: boolean;
+}>`
   &&,
   &&:before {
     display: block;
@@ -73,7 +80,7 @@ const ToolTipArrow = styled.span<{ position?: any; isInverse?: boolean }>`
   }
 `;
 
-const StyledTooltip = styled.div<{
+export const StyledTooltip = styled.div<{
   isInverse?: boolean;
   isVisible?: boolean;
   position: TooltipPosition;
@@ -129,7 +136,7 @@ const StyledTooltip = styled.div<{
 
 // Using any for the ref because it is put ont he passed in children which does not have a specific type
 export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
-  const [isVisible, setIsVisible] = React.useState<boolean>(false);
+  const [isVisible, setIsVisible] = React.useState<boolean>(props.open);
   const [referenceElement, setReferenceElement] =
     React.useState<HTMLElement>(null);
   const [popperElement, setPopperElement] =
@@ -154,7 +161,7 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
   React.useEffect(() => {
     const handleEsc = event => {
       if (event.key === 'Escape') {
-        setIsVisible(false);
+        hideTooltip();
       }
     };
     window.addEventListener('keydown', handleEsc);
@@ -166,7 +173,7 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
 
   function handleKeyDown(event: React.KeyboardEvent) {
     if (event.key === 'Escape') {
-      setIsVisible(false);
+      hideTooltip();
     }
   }
 
@@ -175,7 +182,7 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
   }
 
   function hideTooltip() {
-    setIsVisible(false);
+    setIsVisible(props.open);
   }
 
   const {
@@ -214,7 +221,7 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
   const isInverse = useIsInverse(props.isInverse);
 
   return (
-    <ToolTipContainer
+    <TooltipContainer
       {...other}
       data-testid={testId}
       onKeyDown={handleKeyDown}
@@ -238,7 +245,7 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
             {...attributes.popper}
           >
             {content}
-            <ToolTipArrow
+            <TooltipArrow
               isInverse={isInverse}
               ref={setArrowElement}
               style={combinedArrowStyle}
@@ -247,6 +254,6 @@ export const Tooltip = React.forwardRef<any, TooltipProps>((props, ref) => {
           </StyledTooltip>
         </div>
       )}
-    </ToolTipContainer>
+    </TooltipContainer>
   );
 });